update tests to reflect updated package structure
