+++ artist/core/kinematics_reconstructor.py
@@ -3,13 +3,13 @@ import pathlib
 from typing import Any, cast
 
 import torch
-from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 from torch.optim.lr_scheduler import LRScheduler
 
 from artist.core import core_utils, learning_rate_schedulers
 from artist.core.loss_functions import Loss
 from artist.field.heliostat_group import HeliostatGroup
 from artist.io.calibration_parser import CalibrationDataParser
+from artist.raytracing.heliostat_ray_tracer import HeliostatRayTracer
 from artist.scenario.scenario import Scenario
 from artist.util import config_dictionary, index_mapping, utils
 from artist.util.environment_setup import DdpSetup, get_device
