+++ tests/raytracing/test_blocking.py
@@ -3,10 +3,10 @@ import pathlib
 import h5py
 import pytest
 import torch
-from raytracing import blocking
-from raytracing.heliostat_ray_tracer import HeliostatRayTracer
 
 from artist import ARTIST_ROOT
+from artist.raytracing import blocking
+from artist.raytracing.heliostat_ray_tracer import HeliostatRayTracer
 from artist.scenario.scenario import Scenario
 from artist.util import utils
