+++ tests/raytracing/test_geometry.py
